==> src/1-Dump2CSVs.ipynb <==
!pip install pyspark

!wget http://www.jeuxdemots.org/JDM-LEXICALNET-FR/09032019-LEXICALNET-JEUXDEMOTS-FR-NOHTML.txt.zip
!unzip 09032019-LEXICALNET-JEUXDEMOTS-FR-NOHTML.txt.zip  -d ./
!iconv -f CP1252 -t UTF-8 /content/JDM-LEXICALNET-FR/09032019-LEXICALNET-JEUXDEMOTS-FR-NOHTML.txt -o /content/JDM-LEXICALNET-FR/09032019-LEXICALNET-JEUXDEMOTS-FR-NOHTML-UTF8.txt

import sys
import os
import pyspark as ps
import warnings
import re
import datetime

from pyspark import SparkContext 
from pyspark.sql import SQLContext
from pyspark.sql import SparkSession
from pyspark.sql import functions as F
from pyspark.sql import Row
from pyspark.sql.types import StringType

spark = SparkSession.Builder().appName("RezoJDM-SDS").getOrCreate()
sc = spark.sparkContext

jdmDUMP = sc.textFile("/content/JDM-LEXICALNET-FR/09032019-LEXICALNET-JEUXDEMOTS-FR-NOHTML-UTF8.txt")
# jdmDUMP = sc.textFile("/content/09032019-Template.txt")

jdmNodes = jdmDUMP.filter(lambda x: x[:4] == "eid=")

jdmNodes = jdmNodes.map(lambda x: 
                        Row(
                                eid = int(re.findall("eid=(\d+)\|n=", x)[0]) ,
                                n = re.findall("n=\"(.+)\"\|t=", x)[0] ,
                                t = int(re.findall("t=(-?\d+)\|w=", x)[0]) ,
                                w = int(re.findall("w=(-?\d+)", x)[0])
                            )
                  )

# only type 1 and 2 should be kept
jdmNodes = jdmNodes.filter(lambda x: (x["t"] == 1) | (x["t"] == 2))

# node whose text starts and ends with equal (“=”) are not french, don’t keep them
jdmNodes = jdmNodes.filter(lambda x: (~((x["n"][0] == "=") & (x["n"][-1] == "="))))

# filter the weights lesser than M(>50)
jdmNodes = jdmNodes.filter(lambda x: x["w"] > 50)

dfNodes = jdmNodes.toDF()

jdmNodes.count()

jdmRelations = jdmDUMP.filter(lambda x: x[:4] == "rid=")

jdmRelations = jdmRelations.map(lambda x: 
                                    Row(
                                            rid = int(re.findall("rid=(\d+)\|n1=", x)[0]) ,
                                            n1 = int(re.findall("n1=(\d+)\|n2=", x)[0]) ,
                                            n2 = int(re.findall("n2=(-?\d+)\|t=", x)[0]) ,
                                            t = int(re.findall("t=(-?\d+)\|w=", x)[0]) ,
                                            w = int(re.findall("w=(-?\d+)", x)[0])
                                        )
                            )
# only type 1 and 2 should be kept
jdmRelations = jdmRelations.filter(lambda x: x["w"] > 50)

# keep all nodes except { 12, 18, 29, 33, 36, 45, 46, 47, 48, 1000, 1001, 1002, 444, 128, 200}
black_list_relation_ids = [12, 18, 29, 33, 36, 45, 46, 47, 48, 1000, 1001, 1002, 444, 128, 200]
jdmRelations = jdmRelations.filter(lambda x: x["t"] not in black_list_relation_ids)

# edges might refer to non-existing node, don’t keep those edges
existing_node_ids = list(dfNodes.toPandas()["eid"])
jdmRelations = jdmRelations.filter(lambda x: (x["n1"] in existing_node_ids) & (x["n2"] in existing_node_ids))

jdmRelations.count()

dfNodes = jdmNodes.toDF()
dfNodes = dfNodes.toPandas()

dfRelations = jdmRelations.toDF()
dfRelations = dfRelations.toPandas()

dfNodes.to_csv("nodes.csv", index=False)
dfRelations.to_csv("relations.csv", index=False)

dfNodes

dfRelations

from google.colab import files

files.download("nodes.csv")
files.download("relations.csv")

nodes_out_degree = {}
nodes_in_degree = {}
nodes_degree = {}

for _, row in dfRelations.iterrows():
    
    if row.n1 in nodes_out_degree:
        nodes_out_degree[row.n1]+=1
    else:
        nodes_out_degree[row.n1]=1


    if row.n2 in nodes_in_degree:
        nodes_in_degree[row.n2]+=1
    else:
        nodes_in_degree[row.n2]=1


    if row.n1 in nodes_degree:
        nodes_degree[row.n1]+=1
    else:
        nodes_degree[row.n1]=1


    if row.n2 in nodes_degree:
        nodes_degree[row.n2]+=1
    else:
        nodes_degree[row.n2]=1


nodes_degree

len(dfRelations["t"].unique())

values=[]
keys=[]

for key, value in nodes_degree.items():
    values.append(value)
    keys.append(key)
    if value == 15818:
        print(key)

import numpy as np
min(items), max(items), np.mean(items)

dfNodes[dfNodes["eid"]==116477]
