==> src/2-CSV2Triplets.ipynb <==
import os
import numpy as np
import pandas as pd 
import matplotlib.pyplot as plt
import numpy as np

import networkx as nx
from google.colab import files

! git clone https://github.com/ContentSide/French_Knowledge_Graph.git

nodes = pd.read_csv("/content/French_Knowledge_Graph/datasets/Filtered_Nodes_Edges/nodes.csv", index_col=False)
relations = pd.read_csv("/content/French_Knowledge_Graph/datasets/Filtered_Nodes_Edges/relations.csv", index_col=False)

nodes_out_degree = {}
nodes_in_degree = {}
nodes_degree = {}

for _, row in relations.iterrows():
    
    if row.n1 in nodes_out_degree:
        nodes_out_degree[row.n1]+=1
    else:
        nodes_out_degree[row.n1]=1


    if row.n2 in nodes_in_degree:
        nodes_in_degree[row.n2]+=1
    else:
        nodes_in_degree[row.n2]=1


    if row.n1 in nodes_degree:
        nodes_degree[row.n1]+=1
    else:
        nodes_degree[row.n1]=1


    if row.n2 in nodes_degree:
        nodes_degree[row.n2]+=1
    else:
        nodes_degree[row.n2]=1


values=[]
keys=[]

for key, value in nodes_degree.items():
    values.append(value)
    keys.append(key)
    if value == 15818:
        print(key)

len(relations["t"].unique())

min(values), max(values), np.mean(values)

#Entities connected to less than 45 other nodes (degree(node)<45) were removed.

nodes_to_keep=[]
for key, value in nodes_degree.items():
    if value >= 45:
        nodes_to_keep.append(key)

relations_filter1 = relations[relations["n1"].isin(nodes_to_keep) & relations["n2"].isin(nodes_to_keep)].reset_index()

len(nodes_to_keep)

relations_filter1.shape

len(relations_filter1["t"].unique())

#Relationships that appeared less than 100 times in the graph were also deleted.

relation_freq={}

for _, row in relations_filter1.iterrows():
    
    if row.t in relation_freq:
        relation_freq[row.t]+=1
    else:
        relation_freq[row.t]=1

relations_to_keep=[]
for key, value in relation_freq.items():
    if value >= 100:
        relations_to_keep.append(key)

relations_filter2 = relations_filter1[relations_filter1["t"].isin(relations_to_keep)].reset_index()

relations_filter1.shape, relations_filter2.shape, len(relations_filter1["t"].unique()), len(relations_filter2["t"].unique())

# pairs of symmetric relationship types
symmetic_reltypes = [
  (21,20),
  (53,54),
  (50,51),
  (41,42),
  (31,30),
  (32,115),
  (39,40),
  (14,26),
  (16,25),
  (3,27),
  (73,74),
  (55,56),
  (13,24),
  (6,8),
  (17,23),
  (15,28),
  (1,2000),
  (9,10),
]

_relations_to_delete = set()

error = False


_relations_to_delete = set()

error = False

for n, reltype_pair in enumerate(symmetic_reltypes):
  print("Handling reltype pair: {} ({}/{})\n\n".format(reltype_pair, n+1, len(symmetic_reltypes)))
  _edges = relations_filter2[relations_filter2["t"].isin(reltype_pair)]
  # print("_edges types:")
  # print((_edges["t"].unique()))
  # print(f"type(_edges) = {type(_edges)}\n_edges = {_edges}\n\n")
  for _, _edge in _edges.iterrows():
    # print(f"type(_edge) = {type(_edge)}\n_edge = {_edge}\n")
    # print(f"type(_edge['rid']) = {type(_edge['rid'])}\t_edge['rid'] = {_edge['rid']}\n")
    if _edge["t"] == reltype_pair[0]:
      other_type = reltype_pair[1]
    elif _edge["t"] == reltype_pair[1]:
      other_type = reltype_pair[0]
    # print(f"current edge type: {_edge['t']}\tsymmetric edge expected type: {other_type}")
    symmetric_rel = _edges[(_edges["t"] == other_type) & (_edges["n1"] == _edge["n2"]) & (_edges["n2"] == _edge["n1"])]
    # print(f"type(symmetric_rel) = {type(symmetric_rel)}\nsymmetric_rel = {symmetric_rel}\n")
    # print(f"type(symmetric_rel['rid']) = {type(symmetric_rel['rid'])}\tsymmetric_rel['rid'] = {symmetric_rel['rid']}\n")
    # print(f"type(symmetric_rel['rid'].values[0]) = {type(symmetric_rel['rid'].values[0])}\tsymmetric_rel['rid'].values[0] = {symmetric_rel['rid'].values[0]}\n")
    if symmetric_rel.shape[0] == 0:
      continue
    elif symmetric_rel.shape[0] > 1:
      print(f"Expected 0 or 1 result, got {symmetric_rel.shape[0]} instead")
      print(f"  _edge =\n{_edge}\n")
      print(f"  symmetric_rel =\n{symmetric_rel}")
      print("-" * 15)
    test = (_edge["w"] > symmetric_rel["w"]).values[0]
    _relations_to_delete.add(_edge["rid"] if test else symmetric_rel["rid"].values[0])  
    # print(f"len(_relations_to_delete) = {len(_relations_to_delete)}")

print("number of relations to delete:\n", len(_relations_to_delete))

relations_filter3 = relations_filter2[~relations_filter2["rid"].isin(_relations_to_delete)]
relations_filter2.shape, relations_filter3.shape

relations_filter2.shape[0] - relations_filter3.shape[0]

# Let's make sure every relationship type still has >=100 examples. If not, remove it.

_relation_freq={}

for _, _row in relations_filter3.iterrows():
    
    if _row.t in _relation_freq:
        _relation_freq[_row.t]+=1
    else:
        _relation_freq[_row.t]=1

for _reltype, _relfreq in _relation_freq.items():
  if _relfreq < 100:
    print(f"WARNING! Relationship type {_reltype} has {_relfreq}<100 examples")

relations_to_keep=[]
for key, value in _relation_freq.items():
    if value >= 100:
        relations_to_keep.append(key)

relations_filter4 = relations_filter3[relations_filter3["t"].isin(relations_to_keep)]
relations_filter3.shape, relations_filter4.shape

relations_filter3.shape[0] - relations_filter4.shape[0]

relations_filter2.shape, relations_filter4.shape, len(relations_filter2["t"].unique()), len(relations_filter4["t"].unique())

#code to find components of the graph. The largest subgraph will be selected.

def get_sub_disconnected_graphs(relations):
    G = nx.from_pandas_edgelist(relations, source='n1', target='n2', edge_attr='t')
    largest_component = max(nx.connected_components(G), key=len)
    G2 = G.subgraph(largest_component)
    return len(list(nx.connected_components(G)))

get_sub_disconnected_graphs(relations), get_sub_disconnected_graphs(relations_filter2), get_sub_disconnected_graphs(relations_filter4)

relations = relations_filter4[["n1","n2","t","w"]]
relations.shape , len(relations["t"].unique())

# relations_info = {}

# for _, row in relations.iterrows():
#     start = min(row.n1,row.n2)
#     end = max(row.n1,row.n2)

#     link = str(start) + "-" + str(end)

#     if link in relations_info:
#         relations_info[link].append(row)
#     else:
#         relations_info[link] = [row]

def get_subset_triplet(relations, N):

    relations_ = relations.copy()
    count = 0


    while count < N:

        rand_index = np.random.randint(relations_.shape[0], size=1)[0]
        relations_dropped = relations_.drop(rand_index)

        condition1 = relations_.loc[rand_index]["n1"] in relations_dropped["n1"]
        condition2 = relations_.loc[rand_index]["n2"] in relations_dropped["n2"]    

        if condition1 & condition2:

            if count == 0:
                relations_new = relations_.loc[rand_index].to_frame().T
            else:
                relations_new = relations_new.append(relations_.loc[rand_index], ignore_index=True)

            relations_ = relations_dropped.reset_index()[["n1","n2","t","w"]]
            count+=1

            if count % 1000 == 0:
                print(count)

    return relations_new, relations_

total = 272 + 17 + 20
272/total, 17/total, 20/total

N = 832093
N*0.1 , N*0.1 , N*0.8

triplets_dev, triplets_rest = get_subset_triplet(relations, 83209)
triplets_dev.to_csv("triplets_dev.csv", index=False)
triplets_rest.to_csv("triplets_rest.csv", index=False)
files.download("triplets_dev.csv")
files.download("triplets_rest.csv")

triplets_test, triplets_train = get_subset_triplet(triplets_rest, 83209)
triplets_train.to_csv("triplets_train.csv", index=False)
triplets_test.to_csv("triplets_test.csv", index=False)
files.download("triplets_train.csv")
files.download("triplets_test.csv")
