==> src/4-OpenKE_2_KGE_Models.ipynb <==
#Initiation

!git clone -b OpenKE-PyTorch https://github.com/thunlp/OpenKE
%cd OpenKE/
!mkdir checkpoint
!mkdir result
%cd openke
!bash make.sh
%cd ..
%cd ..
!git clone https://github.com/ContentSide/French_Knowledge_Graph.git

%cd OpenKE/
!cp -r ../French_Knowledge_Graph/benchmarks/RezoJDM16K ./benchmarks

# !cp benchmarks/FB13/n-n.py 'benchmarks/RezoJDM16K'
# %cd 'benchmarks/RezoJDM16K'
# !python n-n.py
# %cd ../..

# No need to run the above lines; since the SDS is ready!
!cp -r ../French_Knowledge_Graph/trainers/*  ./

#List of Models to run

!ls ./*.py

#Run model(s)

!python train_rotate_RezoJDM16K_adv.py

!python train_distmult_RezoJDM16K.py

!python train_complex_RezoJDM16K.py

!python train_transe_RezoJDM16K.py

!python train_transh_RezoJDM16K.py

!python train_transd_RezoJDM16K.py
